Refactor EntityModelAssembler's toCollectionModel method to accept a Function for adding item links. Refactor AirportController to use a Function to add item links for each AirportOutputDTO model in airportCollectionModel. Simplify the getOneAirport method to only set the self-link for the entity model.
This commit refactors the toCollectionModel method of EntityModelAssembler to accept a Function instead of a ToIntFunction for adding item links to the EntityModels. This change allows for more flexibility in how item links are added to the EntityModels. The itemLinks function is applied to each EntityModel in the CollectionModel to add the appropriate links. Additionally, the toModel method now uses the @NonNull annotation to ensure validity. The appropriate changes were also made in the AirportController class